sg_project/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.2
django-heroku==0.0.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-nested-routers==0.93.3
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
psycopg2==2.9.1
pycparser==2.20
pyjwt==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
